Reformat.

+++ mingw_install_deps.sh
@@ -39,7 +39,7 @@ function extract(){
 function printc(){
 	echo -e "\033[33m$1\033[0m"
 }
-#install without hacks
+#install package without hacks
 #$1: URL
 #$2: archive filename to detect http redirection problems.
 function installpackage(){
@@ -91,13 +91,17 @@ fi
 #maybe we should host this file
 #wget -c -P $DLDIR http://download1039.mediafire.com/p2h9ja9uzvtg/g35fh308hmdklz5/rsync-3.0.8.tar.lzma
 #wget -c -P $DLDIR http://k002.kiwi6.com/hotlink/w8nv7zl9qh/rsync_3_0_8_tar.lzma
-mkdir -p rsync
-cd rsync
-wget -c -P $DLDIR http://sajty.elementfx.com/rsync-3.0.8.tar.lzma
-bsdtar -xf $DLDIR/rsync-3.0.8.tar.lzma
-cp rsync.exe $PREFIX/bin/rsync.exe
-cd ..
-
+PKGLOCKFILE="$LOCKDIR/rsync_installed.lock"
+if [ ! -f $PKGLOCKFILE ]; then
+	printc "Installing rsync..."
+	mkdir -p rsync
+	cd rsync
+	wget -q -c -P $DLDIR http://sajty.elementfx.com/rsync-3.0.8.tar.lzma
+	extract $DLDIR/rsync-3.0.8.tar.lzma 2> /dev/null
+	cp rsync.exe $PREFIX/bin/rsync.exe
+	cd ..
+	touch $PKGLOCKFILE
+fi
 #install glib
 FILENAME="glib-2.28.7.tar.bz2"
 installpackage "http://ftp.gnome.org/pub/gnome/sources/glib/2.28/$FILENAME" "$FILENAME"
@@ -109,28 +113,63 @@ FILENAME="pkg-config-0.26.tar.gz"
 installpackage "http://pkgconfig.freedesktop.org/releases/$FILENAME" "$FILENAME"
 
 #install freeimage
-wget -c -P $DLDIR http://downloads.sourceforge.net/freeimage/FreeImage3150.zip
-bsdtar -xf $DLDIR/FreeImage3150.zip
-cd FreeImage
-make -f Makefile.minGW FREEIMAGE_LIBRARY_TYPE=STATIC DLLTOOLFLAGS="" $MAKEOPTS
-cp Dist/FreeImage.a $PREFIX/lib/libFreeImage.a
-cp Dist/FreeImage.h $PREFIX/include/FreeImage.h
-cd ..
+#hacks:
+#	you need to force mingw makefile, or it will use gnu makefile.
+#	Do not set *FLAGS or it will fail.
+#	FREEIMAGE_LIBRARY_TYPE needs to be static or it will make .lib instead of .a file.
+#	DLLTOOLFLAGS needs to be "" or ogre will get linker errors.
+#	it will try to install .lib, so we need to install it manually
+#	it needs python to build documentation. You can't disable documentation.
+PKGNAME="FreeImage"
+PKGLOGDIR="$LOGDIR/$PKGNAME"
+PKGLOCKFILE="$LOCKDIR/${PKGNAME}_installed.lock"
+if [ ! -f $PKGLOCKFILE ]; then
+	printc "Installing $PKGNAME..."
+	mkdir -p $PKGLOGDIR
+	printc "    Downloading..."
+	wget -q -c -P $DLDIR http://downloads.sourceforge.net/freeimage/FreeImage3150.zip
+	printc "    Extracting..."
+	extract $DLDIR/FreeImage3150.zip 2> $PKGLOGDIR/extract.log
+	cd $PKGNAME
+	printc "    Building..."
+	make -f Makefile.minGW FREEIMAGE_LIBRARY_TYPE=STATIC DLLTOOLFLAGS="" > $PKGLOGDIR/build.log
+	printc "    Installing..."
+	cp Dist/FreeImage.a $PREFIX/lib/libFreeImage.a
+	cp Dist/FreeImage.h $PREFIX/include/FreeImage.h
+	cd ..
+	touch $PKGLOCKFILE
+fi
 
+#install cmake
+#hacks:
+#	if you set the *FLAGS environment variables, it may fail on a test.
+#	make install copies to the wrong location
+PKGNAME="cmake-2.8.3"
+PKGLOGDIR="$LOGDIR/$PKGNAME"
+PKGLOCKFILE="$LOCKDIR/${PKGNAME}_installed.lock"
+if [ ! -f $PKGLOCKFILE ]; then
+	printc "Installing $PKGNAME..."
+	mkdir -p $PKGLOGDIR
+	printc "    Downloading..."
+	wget -q -c -P $DLDIR http://www.cmake.org/files/v2.8/$PKGNAME.tar.gz
+	printc "    Extracting..."
+	extract $DLDIR/$PKGNAME.tar.gz 2> $PKGLOGDIR/extract.log
+	mkdir -p $PKGNAME/mingw_build
+	cd $PKGNAME/mingw_build
+	printc "    Running configure..."
+	../configure --prefix=$PREFIX $CONFIGURE_EXTRA_FLAGS > $PKGLOGDIR/configure.log
+	printc "    Building..."
+	make $MAKEOPTS > $PKGLOGDIR/build.log
+	printc "    Installing..."
+	cp bin/cmake.exe $PREFIX/bin/cmake.exe
+	make install > $PKGLOGDIR/install.log
+	cd ../..
+	touch $PKGLOCKFILE
+fi
 
 export CFLAGS="-O3 -msse2 -ffast-math -mthreads -DNDEBUG -I$PREFIX/include $CFLAGS"
 #without -msse2 ogre is not building
 export CXXFLAGS="-O3 -msse2 -ffast-math -mthreads -DNDEBUG -I$PREFIX/include $CXXFLAGS"
-
-#install cmake
-wget -c -P $DLDIR http://www.cmake.org/files/v2.8/cmake-2.8.4.tar.gz
-bsdtar -xf $DLDIR/cmake-2.8.4.tar.gz
-cd cmake-2.8.4
-./configure --prefix=$PREFIX
-make $MAKEOPTS
-cp bin/cmake.exe $PREFIX/bin/cmake.exe
-make install
-cd ..
 
 #install zziplib
 FILENAME="zziplib-0.13.60.tar.bz2"
@@ -178,100 +217,138 @@ if [ ! -f $PKGLOCKFILE ]; then
 fi
 
 #install boost
-#bjam generated in msys is not working, we need prebuilt bjam.
-wget -c -P $DLDIR http://sourceforge.net/projects/boost/files/boost-jam/3.1.18/boost-jam-3.1.18-1-ntx86.zip/download
-bsdtar -xf $DLDIR/boost-jam-3.1.18-1-ntx86.zip
-wget -c -P $DLDIR http://sourceforge.net/projects/boost/files/boost/1.46.1/boost_1_46_1.tar.bz2/download
-bsdtar -xf $DLDIR/boost_1_46_1.tar.bz2
-ln -s $PWD/boost-jam-3.1.18-1-ntx86/bjam.exe $PWD/boost_1_46_1/bjam
-cd boost_1_46_1
-./bjam --with-thread --with-date_time --stagedir=$PREFIX --layout=system variant=release link=shared toolset=gcc
-#"./bjam install" is not working in msys, we need to install the headers manually.
-cp -r boost $PREFIX/include
-cd ..
-mv $PREFIX/lib/libboost_date_time.dll $PREFIX/bin/libboost_date_time.dll
-mv $PREFIX/lib/libboost_thread.dll $PREFIX/bin/libboost_thread.dll
+#hacks:
+#	bjam generated in msys is not working, we need prebuilt bjam.
+#	"./bjam install" is not working in msys, we need to install it manually.
+PKGLOCKFILE="$LOCKDIR/boost_installed.lock"
+if [ ! -f $PKGLOCKFILE ]; then
+	wget -c -P $DLDIR http://sourceforge.net/projects/boost/files/boost-jam/3.1.18/boost-jam-3.1.18-1-ntx86.zip/download
+	extract $DLDIR/boost-jam-3.1.18-1-ntx86.zip
+	wget -c -P $DLDIR http://sourceforge.net/projects/boost/files/boost/1.46.1/boost_1_46_1.tar.bz2/download
+	extract $DLDIR/boost_1_46_1.tar.bz2
+	ln -s $PWD/boost-jam-3.1.18-1-ntx86/bjam.exe $PWD/boost_1_46_1/bjam
+	cd boost_1_46_1
+	./bjam --with-thread --with-date_time --stagedir=$PREFIX --layout=system variant=release link=shared toolset=gcc
+	#"./bjam install" is not working in msys, we need to install the headers manually.
+	cp -r boost $PREFIX/include
+	mv $PREFIX/lib/libboost_date_time.dll $PREFIX/bin/libboost_date_time.dll
+	mv $PREFIX/lib/libboost_thread.dll $PREFIX/bin/libboost_thread.dll
+	cd ..
+	touch $PKGLOCKFILE
+fi
 
 #install lua
-wget -c -P $DLDIR http://www.lua.org/ftp/lua-5.1.4.tar.gz
-bsdtar -xf $DLDIR/lua-5.1.4.tar.gz
-cd lua-5.1.4
-make mingw $MAKEOPTS
-make install INSTALL_TOP=$PREFIX
-cd ..
-
+PKGLOCKFILE="$LOCKDIR/lua_installed.lock"
+if [ ! -f $PKGLOCKFILE ]; then
+	wget -c -P $DLDIR http://www.lua.org/ftp/lua-5.1.4.tar.gz
+	extract $DLDIR/lua-5.1.4.tar.gz
+	cd lua-5.1.4
+	make mingw $MAKEOPTS
+	make install INSTALL_TOP=$PREFIX
+	#install lua5.1.pc
+	wget -c -P $DLDIR http://sajty.elementfx.com/lua5.1.pc
+	cp $DLDIR/lua5.1.pc ./lua5.1.pc
+	export PREFIX_ESCAPED=$(echo $PREFIX | sed -e 's/\(\/\|\\\|&\)/\\&/g')
+	sed -i "s/TPL_PREFIX/$PREFIX_ESCAPED/g" ./lua5.1.pc 
+	mv ./lua5.1.pc $PREFIX/lib/pkgconfig/lua5.1.pc
+	cd ..
+	touch $PKGLOCKFILE
+fi
 #install tolua++
-wget -c -P $DLDIR http://www.codenix.com/~tolua/tolua++-1.0.93.tar.bz2
-bsdtar -xf $DLDIR/tolua++-1.0.93.tar.bz2
-cd tolua++-1.0.93
-#tolua uses scons, which needs python, which is big.
-#note: creating makefile would be better
-cp include/tolua++.h $PREFIX/include/tolua++.h
-cd src/lib
-gcc $CFLAGS -c -I$PREFIX/include -L$PREFIX/lib *.c
-ar cq libtolua++.a *.o
-cp libtolua++.a $PREFIX/lib/libtolua++.a
-cd ../bin
-gcc $CFLAGS $LDFLAGS -o tolua++ -I$PREFIX/include -L$PREFIX/lib -mwindows tolua.c toluabind.c -ltolua++ -llua
-cp tolua++.exe $PREFIX/bin/tolua++.exe
-cd ../../..
+#hacks:
+#	tolua uses scons, which needs python, which is big.
+PKGLOCKFILE="$LOCKDIR/tolua++_installed.lock"
+if [ ! -f $PKGLOCKFILE ]; then
+	wget -c -P $DLDIR http://www.codenix.com/~tolua/tolua++-1.0.93.tar.bz2
+	extract $DLDIR/tolua++-1.0.93.tar.bz2
+	cd tolua++-1.0.93
+	cp include/tolua++.h $PREFIX/include/tolua++.h
+	cd src/lib
+	gcc $CFLAGS -c -I$PREFIX/include -L$PREFIX/lib *.c
+	ar cq libtolua++.a *.o
+	cp libtolua++.a $PREFIX/lib/libtolua++.a
+	cd ../bin
+	gcc $CFLAGS $LDFLAGS -o tolua++ -I$PREFIX/include -L$PREFIX/lib -mwindows tolua.c toluabind.c -ltolua++ -llua
+	cp tolua++.exe $PREFIX/bin/tolua++.exe
+	cd ../../..
+	touch $PKGLOCKFILE
+fi
 
 #install openal-soft
-wget -c -P $DLDIR http://kcat.strangesoft.net/openal-releases/openal-soft-1.13.tar.bz2
-bsdtar -xf $DLDIR/openal-soft-1.13.tar.bz2
-cd openal-soft-1.13
-cd build
-cmake -DCMAKE_INSTALL_PREFIX=$PREFIX -G"MSYS Makefiles" ..
-make all $MAKEOPTS
-make install
-cd ../..
+PKGLOCKFILE="$LOCKDIR/openal-soft_installed.lock"
+if [ ! -f $PKGLOCKFILE ]; then
+	wget -c -P $DLDIR http://kcat.strangesoft.net/openal-releases/openal-soft-1.13.tar.bz2
+	extract $DLDIR/openal-soft-1.13.tar.bz2
+	cd openal-soft-1.13
+	cd build
+	cmake -DCMAKE_INSTALL_PREFIX=$PREFIX -G"MSYS Makefiles" ..
+	make all $MAKEOPTS
+	make install
+	cd ../..
+	touch $PKGLOCKFILE
+fi
 
 #install freealut
-wget -c -P $DLDIR http://connect.creativelabs.com/openal/Downloads/ALUT/freealut-1.1.0-src.zip
-bsdtar -xf $DLDIR/freealut-1.1.0-src.zip
-cd freealut-1.1.0-src
-./autogen.sh
-./configure --prefix=$PREFIX $CONFIGURE_EXTRA_FLAGS
-make CFLAGS="$CFLAGS `pkg-config --cflags openal`" LDFLAGS="$LDFLAGS `pkg-config --libs openal`" $MAKEOPTS
-make install
-cd ..
+FILENAME="freealut-1.1.0-src.zip"
+installpackage "http://connect.creativelabs.com/openal/Downloads/ALUT/$FILENAME" "$FILENAME"
 
-#install CG
-cp -r "$CG_INC_PATH" $PREFIX
-cp "$CG_BIN_PATH/cg.dll" $PREFIX/bin/cg.dll
-pexports $PREFIX/bin/cg.dll | sed "s/^_//" > libCg.def
-dlltool --add-underscore -d libCg.def -l $PREFIX/lib/libCg.a
-rm libCg.def
+#install Cg
+PKGLOCKFILE="$LOCKDIR/Cg_installed.lock"
+if [ ! -f $PKGLOCKFILE ]; then
+	cp -r "$CG_INC_PATH" $PREFIX
+	cp "$CG_BIN_PATH/cg.dll" $PREFIX/bin/cg.dll
+	pexports $PREFIX/bin/cg.dll | sed "s/^_//" > libCg.def
+	dlltool --add-underscore -d libCg.def -l $PREFIX/lib/libCg.a
+	rm libCg.def
+	touch $PKGLOCKFILE
+fi
+
 
 #install Ogre3D
-wget -c -P $DLDIR http://sourceforge.net/projects/ogre/files/ogre/1.7/ogre_src_v1-7-3.tar.bz2/download
-7za x -y $DLDIR/ogre_src_v1-7-3.tar.bz2
-7za x -y ogre_src_v1-7-3.tar
-rm ogre_src_v1-7-3.tar
-cd ogre_src_v1-7-3
-mkdir -p build
-cd build
-cmake --with-gui=win32 --with-platform=win32 --enable-direct3d -DCMAKE_INSTALL_PREFIX=$PREFIX -G"MSYS Makefiles" -DOGRE_INSTALL_DEPENDENCIES=false \
--DOGRE_BUILD_SAMPLES=false -DOGRE_BUILD_PLUGIN_BSP=false -DOGRE_BUILD_PLUGIN_OCTREE=false -DOGRE_BUILD_PLUGIN_PCZ=false -DOGRE_BUILD_COMPONENT_RTSHADERSYSTEM=false ..
-make all $MAKEOPTS
-make install
-cp -r $PREFIX/bin/RelWithDebInfo/* $PREFIX/bin
-cp -r $PREFIX/lib/RelWithDebInfo/* $PREFIX/lib
-cd ../..
+#hacks:
+#	its not creating ogre.pc, we need to create them manually.
+PKGLOCKFILE="$LOCKDIR/Ogre_installed.lock"
+if [ ! -f $PKGLOCKFILE ]; then
+	wget -c -P $DLDIR http://sourceforge.net/projects/ogre/files/ogre/1.7/ogre_src_v1-7-3.tar.bz2/download
+	extract $DLDIR/ogre_src_v1-7-3.tar.bz2
+	cd ogre_src_v1-7-3
+	mkdir -p build
+	cd build
+	cmake -DCMAKE_INSTALL_PREFIX=$PREFIX -G"MSYS Makefiles" \
+	-DOGRE_INSTALL_DEPENDENCIES=false -DOGRE_BUILD_SAMPLES=false -DOGRE_BUILD_PLUGIN_BSP=false \
+	-DOGRE_BUILD_PLUGIN_OCTREE=false -DOGRE_BUILD_PLUGIN_PCZ=false -DOGRE_BUILD_RENDERSYSTEM_D3D9=false \
+	-DOGRE_BUILD_COMPONENT_RTSHADERSYSTEM=false ..
+	make all $MAKEOPTS
+	make install
+	#copy to get it available without configuration name.
+	cp -r $PREFIX/bin/RelWithDebInfo/* $PREFIX/bin
+	cp -r $PREFIX/lib/RelWithDebInfo/* $PREFIX/lib
+
+	#get *.pc files.
+	wget -c -P $DLDIR http://sajty.elementfx.com/ogre_package.zip
+	extract $DLDIR/ogre_package.zip
+	export PREFIX_ESCAPED=$(echo $PREFIX | sed -e 's/\(\/\|\\\|&\)/\\&/g')
+	find . -maxdepth 1 -name "OGRE*.pc" -exec sed -i "s/TPL_PREFIX/$PREFIX_ESCAPED/g" '{}' \;
+	mv ./OGRE*.pc $PREFIX/lib/pkgconfig
+	cd ../..
+	touch $PKGLOCKFILE
+fi
 
 #install CEGUI
-wget -c -P $DLDIR http://sourceforge.net/projects/crayzedsgui/files/CEGUI%20Mk-2/0.7.5/CEGUI-0.7.5.tar.gz/download
-bsdtar -xf $DLDIR/CEGUI-0.7.5.tar.gz
-cd CEGUI-0.7.5
-./configure --prefix=$PREFIX --disable-samples --disable-opengl-renderer --disable-irrlicht-renderer --disable-xerces-c --disable-libxml --disable-expat --disable-directfb-renderer \
---enable-freeimage --enable-ogre-renderer --enable-lua-module --enable-external-toluapp \
-Ogre_CFLAGS="-I$PREFIX/include -I$PREFIX/include/OGRE" Ogre_LIBS="-lOgreMain" \
-FreeImage_CFLAGS="-DUSE_FREEIMAGE_LIBRARY -I$PREFIX/include" FreeImage_LIBS="-lFreeImage" \
-Lua_CFLAGS="-I$PREFIX/include" Lua_LIBS="-llua" \
-toluapp_CFLAGS="-I$PREFIX/include" toluapp_LIBS="-ltolua++" \
-MINGW32_BUILD=true CEGUI_BUILD_LUA_MODULE_UNSAFE=true CEGUI_BUILD_TOLUAPPLIB=true \
-$CONFIGURE_EXTRA_FLAGS
-make $MAKEOPTS
-make install
-cd ..
-
+PKGLOCKFILE="$LOCKDIR/CEGUI_installed.lock"
+if [ ! -f $PKGLOCKFILE ]; then
+	wget -c -P $DLDIR http://sourceforge.net/projects/crayzedsgui/files/CEGUI%20Mk-2/0.7.5/CEGUI-0.7.5.tar.gz/download
+	extract $DLDIR/CEGUI-0.7.5.tar.gz
+	cd CEGUI-0.7.5
+	./configure --prefix=$PREFIX --disable-samples --disable-opengl-renderer --disable-irrlicht-renderer --disable-xerces-c \
+	--disable-libxml --disable-expat --disable-directfb-renderer \
+	--enable-freeimage --enable-ogre-renderer --enable-lua-module --enable-external-toluapp \
+	FreeImage_CFLAGS="-DUSE_FREEIMAGE_LIBRARY -I$PREFIX/include" FreeImage_LIBS="-lFreeImage" \
+	toluapp_CFLAGS="-I$PREFIX/include" toluapp_LIBS="-ltolua++" \
+	MINGW32_BUILD=true CEGUI_BUILD_LUA_MODULE_UNSAFE=true CEGUI_BUILD_TOLUAPPLIB=true \
+	$CONFIGURE_EXTRA_FLAGS
+	make $MAKEOPTS
+	make install
+	cd ..
+	touch $PKGLOCKFILE
+fi
